+++ app/components/features/vatsim/controllers/VatsimControllerInfo.vue
@@ -233,16 +233,10 @@ const isCopied = (key: string) => {
         align-items: center;
 
         overflow-wrap: anywhere;
+    }
 
-        @include mobileOnly{
-            row-gap: 0;
-
-            > .__spacer {
-                position: relative;
-                width: 100%;
-                margin-bottom: 8px;
-            }
-        }
+    &_name {
+        color: var(--color);
     }
 
     &_frequency {